+++ src/cc/hooks/useClusterLoadingState.js
@@ -1,5 +1,4 @@
 import { useConfig } from '../store/ConfigProvider';
-import { useBasicAuth } from '../store/BasicAuthProvider';
 import { useSsoAuth } from '../store/SsoAuthProvider';
 import { useClusterData } from '../store/ClusterDataProvider';
 import { useClusterActions } from '../store/ClusterActionsProvider';
@@ -21,14 +20,6 @@ export const useClusterLoadingState = function () {
 
   const {
     state: {
-      loading: basicAuthLoading,
-      loaded: basicAuthLoaded,
-      error: basicAuthError,
-    },
-  } = useBasicAuth();
-
-  const {
-    state: {
       loading: ssoAuthLoading,
       loaded: ssoAuthLoaded,
       error: ssoAuthError,
@@ -46,15 +37,11 @@ export const useClusterLoadingState = function () {
   return (
     // loading config for cloudUrl
     configLoading ||
-    // performing basic auth
-    basicAuthLoading ||
-    // since we always auto-load clusters once auth is done, fill a potential
-    //  gap that would flash loading state `true > false > true` between end
-    //  of auth load and beginning if cluster load
-    (basicAuthLoaded && !basicAuthError && !clusterDataLoaded) ||
     // performing SSO auth
     ssoAuthLoading ||
-    // fill the same auth > clusters gap as in basic auth case above
+    // since we always auto-load clusters once auth is done, fill a potential
+    //  gap that would flash loading state `true > false > true` between end
+    //  of auth load and beginning of cluster load
     (ssoAuthLoaded && !ssoAuthError && !clusterDataLoaded) ||
     // loading list of clusters
     clusterDataLoading ||